Move printing of messages to the tool
Instead of printing the warnings in the checkMkDerivationFor function in Nix
using builtins.trace, let’s pass the messages through an attribute and
extract them as a JSON object during evaluation triggered by the tool.

This will rid us of the ugly “trace: ” prefix and also allow us to more
easily control which overlays are used.