python-include-tests: fix errors when setuptools-check-hook is used
no-python-tests also fired when a package uses setuptools-check-hook,
which is the default for format = "setuptools".

This change also makes the test more robust. Checks in Python packages
are actually install checks, however buildPythonPackage also accepts
doCheck/checkInputs, which are then used to set
doInstallCheck/installCheckInputs. Consequently, it is also valid for
a derivation to use doInstallCheck/installCheckInputs instead.

To complicate matters, we cannot just check derivation arguments,
since buildPythonPackage also sets doInstallCheck/installCheckInputs
when the project format is setuptools.

This change updates the test to look for check hooks in the
nativeBuildInputs of the final derivation. As well as checking
whether install checks are enabled.