README: Fix stable Nix instructions
For a long time Nixpkgs `nixVersions.stable` was stuck on Nix 2.3.
That version supported `nix run` as a nicer version of `nix-shell`.
In later versions, the command was reclassified as an experimental feature.
Additionally, semantics was changed.

Nixpkgs has bumped stable Nix to a newer version a while ago,
so we cannot continue to use `nix run`.
Let’s replace the legacy command with venerable `nix-shell`.

This required creating a new shell that contains the `nixpkgs-hammering` package.